api/pyproject.toml
[project]
name = "peer-fitness-buddy-api"
version = "0.1.0"
description = "Backend service for the Peer Fitness Buddy App"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "alembic",
    "argon2-cffi",
    "email-validator",
    "gunicorn",
    "psycopg[binary]",
    "python-dotenv",
    "SQLAlchemy",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.12.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "pfb_api"

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["B008"]

[tool.pytest.ini_options]
pythonpath = ["src"]